email_scraper/requirements.txt
requests
python-dotenv
beautifulsoup4
scrapy
pandas